+++ app/src/main/java/com/eazyrento/customer/notification/model/NotificationModel.kt
@@ -3,14 +3,14 @@ package com.eazyrento.customer.notification.model
 
 abstract class BaseNoticationModel {
     abstract val deep_link: String?
-    abstract val id: Long
+    abstract val id: Int
     abstract val obj_id:String
 }
 
 data class NotificationModel(
 
     override val deep_link: String?,
-    override val id: Long,
+    override val id: Int,
     override val obj_id: String,
     val sent_message:String,
     val requested_amount:Double,
